% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{check_renaissance}
\alias{check_renaissance}
\title{Check whether data resembles typical measurement results.}
\usage{
check_renaissance(.data)
}
\arguments{
\item{.data}{Measurement results to test.}
}
\value{
Error message or TRUE.
}
\description{
Check whether data resembles typical measurement results.
}
\details{
Performs basic sanity checking on data,
including presence of essential columns.
}
