% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{compute_flat_percentile_ci}
\alias{compute_flat_percentile_ci}
\title{Compute flat percentile bootstrap confidence interval.}
\usage{
compute_flat_percentile_ci(.input, .column, ...)
}
\arguments{
\item{.input}{Data not structured into runs.}

\item{.column}{Column to compute confidence interval for.}

\item{...}{Parameters to \code{\link[=compute_vector_flat_percentile_ci]{compute_vector_flat_percentile_ci()}}.}
}
\value{
Summarized tibble with confidence interval columns.
}
\description{
Uses \code{\link[=compute_vector_flat_percentile_ci]{compute_vector_flat_percentile_ci()}} to compute percentile confidence interval on data that is not structured into runs.
}
