% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{compute_flat_percentile_cis}
\alias{compute_flat_percentile_cis}
\title{Compute flat percentile bootstrap confidence interval.}
\usage{
compute_flat_percentile_cis(
  .data,
  .column,
  .statistic = mean,
  .confidence = 0.99,
  .R = 10000
)
}
\arguments{
\item{.data}{Data summarized across runs.}

\item{.column}{Column to compute confidence interval for.}

\item{.statistic}{Statistic to compute confidence interval for.}

\item{.confidence}{Confidence level.}

\item{.R}{Replicates.}
}
\value{
Summarized tibble with confidence interval columns.
}
\description{
Uses bootstrap to compute percentile confidence interval on data summarized across runs.
}
