% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{compute_flat_percentile_cis_group_helper}
\alias{compute_flat_percentile_cis_group_helper}
\title{Group helper for flat percentile bootstrap confidence interval.}
\usage{
compute_flat_percentile_cis_group_helper(.input, .statistic, .confidence, .R)
}
\arguments{
\item{.input}{Data to compute confidence interval for.}

\item{.statistic}{Statistic to compute confidence interval for.}

\item{.confidence}{Confidence level.}

\item{.R}{Number of replicates.}
}
\value{
Tibble with confidence interval.
}
\description{
Group helper for flat percentile bootstrap confidence interval.
}
