% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{compute_hierarchical_relative_percentile_ci}
\alias{compute_hierarchical_relative_percentile_ci}
\title{Compute hierarchical relative percentile bootstrap confidence interval.}
\usage{
compute_hierarchical_relative_percentile_ci(.input, .baseline, .column, ...)
}
\arguments{
\item{.input}{Data structured into runs.}

\item{.baseline}{Baseline virtual machine.}

\item{.column}{Column to compute confidence interval for.}

\item{...}{Parameters to \code{\link[=compute_list_vectors_pair_hierarchical_percentile_ci]{compute_list_vectors_pair_hierarchical_percentile_ci()}}.}
}
\value{
Summarized tibble with confidence interval columns.
}
\description{
Uses \code{\link[=compute_list_vectors_pair_hierarchical_percentile_ci]{compute_list_vectors_pair_hierarchical_percentile_ci()}} to compute percentile
confidence interval on data that is structured into runs.
}
\details{
The statistic is computed relative to baseline virtual machine.
}
