% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{compute_list_vectors_hierarchical_percentile_ci}
\alias{compute_list_vectors_hierarchical_percentile_ci}
\title{Compute hierarchical percentile bootstrap confidence interval.}
\usage{
compute_list_vectors_hierarchical_percentile_ci(
  .input,
  .statistic_outer = mean,
  .statistic_inner = mean,
  .confidence = 0.99,
  .R = 10000,
  .reduction = 0
)
}
\arguments{
\item{.input}{Data to compute confidence interval for.}

\item{.statistic_outer}{Statistic to apply at the outer level.}

\item{.statistic_inner}{Statistic to apply at the inner leve.}

\item{.confidence}{Confidence level.}

\item{.R}{Number of replicates.}

\item{.reduction}{Replicate size reduction.}
}
\value{
Tibble with confidence interval.
}
\description{
Compute hierarchical percentile bootstrap confidence interval.
}
