% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{compute_list_vectors_pair_hierarchical_percentile_ci}
\alias{compute_list_vectors_pair_hierarchical_percentile_ci}
\title{Compute hierarchical percentile bootstrap confidence interval on data set pair.}
\usage{
compute_list_vectors_pair_hierarchical_percentile_ci(
  .input_one,
  .input_two,
  .combination = .Primitive("-"),
  .statistic_outer = mean,
  .statistic_inner = mean,
  .confidence = 0.99,
  .R = 10000,
  .reduction = 0
)
}
\arguments{
\item{.input_one}{Data set one to compute with.}

\item{.input_two}{Data set two to compute with.}

\item{.combination}{Combination to compute for both data sets.}

\item{.statistic_outer}{Statistic to compute at the outer level.}

\item{.statistic_inner}{Statistic to compute at the inner leve.}

\item{.confidence}{Confidence level.}

\item{.R}{Number of replicates.}

\item{.reduction}{Replicate size reduction.}
}
\value{
Tibble with confidence interval.
}
\description{
Compute hierarchical percentile bootstrap confidence interval on data set pair.
}
