% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{compute_trends}
\alias{compute_trends}
\title{Compute trends.}
\usage{
compute_trends(.input, .column)
}
\arguments{
\item{.input}{Data.}

\item{.column}{Column to compute trend for.}
}
\value{
Summarized tibble with trends column.
}
\description{
Uses linear least squares regression to compute trend of each run.
}
