% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{compute_vector_flat_percentile_ci}
\alias{compute_vector_flat_percentile_ci}
\title{Compute flat percentile bootstrap confidence interval.}
\usage{
compute_vector_flat_percentile_ci(
  .input,
  .statistic = mean,
  .confidence = 0.99,
  .R = 10000
)
}
\arguments{
\item{.input}{Data to compute confidence interval for.}

\item{.statistic}{Statistic to compute confidence interval for.}

\item{.confidence}{Confidence level.}

\item{.R}{Number of replicates.}
}
\value{
Tibble with confidence interval.
}
\description{
Compute flat percentile bootstrap confidence interval.
}
