% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{expect_renaissance}
\alias{expect_renaissance}
\title{Check whether data resembles typical measurement results.}
\usage{
expect_renaissance(.data, info = NULL, label = vname(x))
}
\arguments{
\item{.data}{Measurement results to test.}

\item{info}{Internal.}

\item{label}{Internal.}
}
\value{
An expectation.
}
\description{
Check whether data resembles typical measurement results.
}
\seealso{
\code{\link[=check_renaissance]{check_renaissance()}}
}
