% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{flag_outliers_global}
\alias{flag_outliers_global}
\title{Flag outliers.}
\usage{
flag_outliers_global(.input, .column, .flag, ...)
}
\arguments{
\item{.input}{Data.}

\item{.column}{Column to identify outliers in.}

\item{.flag}{Column to use as flag.}

\item{...}{Parameters to \code{\link[=identify_vector_outliers_global]{identify_vector_outliers_global()}}.}
}
\value{
Tibble with new flag column.
}
\description{
Uses \code{\link[=identify_vector_outliers_global]{identify_vector_outliers_global()}} to identify outliers in given column. Then, flags the outlier rows.
}
