% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{identify_vector_outliers_global}
\alias{identify_vector_outliers_global}
\title{Compute outlier flags.}
\usage{
identify_vector_outliers_global(.data, .limit = 0.05, .slack = 0.1)
}
\arguments{
\item{.data}{Sample vector.}

\item{.limit}{Quantile that separates inliers from outliers.}

\item{.slack}{Tolerated distance from limit quantile.}
}
\value{
Boolean vector of outlier flags.
}
\description{
The outlier detection method is based on quantiles but permits retaining short tails.
A sample is considered an outlier if it is outside an interval computed as
\verb{(lo - range * slack, hi + range * slack)} with \code{lo} and \code{hi} denoting
the \code{limit} and \code{1 - limit} quantiles and \code{range = hi - lo}.
}
\examples{
# An outlier gets flagged.
identify_outliers_global (c (1:10, 100))

# No outlier gets flagged with short tails.
identify_outliers_global (1:10)

}
\seealso{
\code{\link[=identify_outliers_window]{identify_outliers_window()}}
}
