% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{identify_vector_outliers_window}
\alias{identify_vector_outliers_window}
\title{Compute outlier flags using sliding window.}
\usage{
identify_vector_outliers_window(
  .input,
  .window = 333,
  .limit = 0.05,
  .slack = 0.1
)
}
\arguments{
\item{.input}{Sample vector.}

\item{.window}{Window size in samples.}

\item{.limit}{Quantile that separates inliers from outliers.}

\item{.slack}{Tolerated distance from limit quantile.}
}
\value{
Boolean vector of outlier flags.
}
\description{
Same as \code{\link[=identify_vector_outliers_global]{identify_vector_outliers_global()}} except the quantile computation is performed
in a sliding window centered on each sample. Currently the computation is not
really efficient and runs in \code{O (n*w*log (w))} for window size \code{w}.
}
\examples{
# A sample of 100 is considered an outlier when near
# different values but not when near similar values.
identify_vector_outliers_window (c (100, 15:25, 95:105, 100), .window = 10)

}
\seealso{
\code{\link[=identify_vector_outliers_global]{identify_vector_outliers_global()}}
}
