% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular.R
\name{list_dimensions}
\alias{list_dimensions}
\title{Get dimensions.}
\usage{
list_dimensions(.input)
}
\arguments{
\item{.input}{Data.}
}
\value{
Tibble with data dimensions.
}
\description{
Returns dimensions of the data, useful to compute relative frequencies in regular position identification.
}
