% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{list_outliers_group_helper}
\alias{list_outliers_group_helper}
\title{Group helper for list outliers.}
\usage{
list_outliers_group_helper(.flags, .index, .total)
}
\arguments{
\item{.flags}{Outlier flags.}

\item{.index}{Data index.}

\item{.total}{Data total.}
}
\value{
Tibble with outlier list.
}
\description{
Given the outlier flags, the index and the total time,
returns a list of outlier positions and intervals.
}
\details{
The computation assumes that an outlier impacts a single repetition,
the outlier interval is therefore the interval of that repetition.
}
