% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{list_outliers_window}
\alias{list_outliers_window}
\title{List outliers using sliding window.}
\usage{
list_outliers_window(.data, .column, ...)
}
\arguments{
\item{.data}{Data.}

\item{.column}{Column to identify outliers in.}

\item{...}{Parameters to \code{\link[=identify_vector_outliers_window]{identify_vector_outliers_window()}}.}
}
\value{
Tibble with outlier list.
}
\description{
Uses \code{\link[=identify_vector_outliers_window]{identify_vector_outliers_window()}} to identify outliers in given column. Then, returns the list of outlier positions and intervals.
}
