% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular.R
\name{list_regular_artifacts_by_interval}
\alias{list_regular_artifacts_by_interval}
\title{Identify regular occurrences based on intervals.}
\usage{
list_regular_artifacts_by_interval(
  .artifacts,
  .dimensions,
  .min_share = 2/3,
  .min_runs = 11
)
}
\arguments{
\item{.artifacts}{Artifact positions and intervals.}

\item{.dimensions}{Data dimensions to compute share from.}

\item{.min_share}{Minimum share of runs to identify regular artifacts.}

\item{.min_runs}{Minimum count of runs to compute share from.}
}
\value{
Tibble with regular artifacts list.
}
\description{
Identifies regular occurrences in a list of artifact positions and intervals.
An artifact is considered regular if it appears in the same interval
in certain minimum share of runs.
}
