% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{list_segment_boundaries}
\alias{list_segment_boundaries}
\title{List segment boundaries.}
\usage{
list_segment_boundaries(.input, .column, ...)
}
\arguments{
\item{.input}{Data.}

\item{.column}{Column to identify segments in.}

\item{...}{Parameters to \code{\link[=locate_vector_segments]{locate_vector_segments()}}.}
}
\value{
Tibble with segment boundary list.
}
\description{
Uses \code{\link[=locate_vector_segments]{locate_vector_segments()}} to identify segments in given column. Then, returns the list of segment boundary positions and intervals.
}
