% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{list_segment_boundaries_group_helper}
\alias{list_segment_boundaries_group_helper}
\title{Group helper for list segment boundaries.}
\usage{
list_segment_boundaries_group_helper(.positions, .index, .total)
}
\arguments{
\item{.positions}{Segment positions.}

\item{.index}{Data index.}

\item{.total}{Data total.}
}
\value{
Tibble with segment boundary list.
}
\description{
Given the segment positions, the index and the total time,
returns a list of segment boundary positions and intervals.
}
\details{
The computation assumes that a segment boundary exists in either boundary repetition,
the boundary interval is therefore formed by the mid points of the two repetitions,
arguing that less than half of each boundary repetition is impacted
(otherwise segments would be assigned differently).

Head and tail boundaries are not included.
}
