% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_file_json}
\alias{load_file_json}
\title{Load data from given file.}
\usage{
load_file_json(data_file)
}
\arguments{
\item{data_file}{Data file to load from.}
}
\value{
Tibble with loaded data.
}
\description{
Load data from given file.
}
\details{
The basic timing columns are translated into canonical form:
\itemize{
\item \code{time} will contain individual repetition time in seconds
\item \code{total} will contain aggregate execution time at the end of repetition in seconds
\item \code{index} will contain the index of the measurement in the run starting from 1
}

Additional metadata is included as factor columns:
\itemize{
\item \code{vm} is a hash of all configuration metadata and can be used to distinguish incompatible measurement configurations
\item \code{vm_name} is the name of the virtual machine
\item \code{vm_version} is the version of the virtual machine
\item \code{vm_configuration} gives concatenated command line arguments
\item \code{run} is a unique identifier of the run
\item \code{benchmark} is the name of the benchmark
}
}
\examples{
load_file_json (rren_example ('results-small-version-5.json'))

}
\seealso{
\code{\link[=load_path_json]{load_path_json()}}
}
