% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_path_json}
\alias{load_path_json}
\title{Load data from given path recursively.}
\usage{
load_path_json(data_path)
}
\arguments{
\item{data_path}{data path to load from}
}
\value{
tibble with loaded data
}
\description{
Load data from given path recursively.
}
\seealso{
\code{\link[=load_file_json]{load_file_json()}}
}
