% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{locate_vector_segments}
\alias{locate_vector_segments}
\title{Locate change point segments in samples.}
\usage{
locate_vector_segments(.data, .penalty)
}
\arguments{
\item{.data}{Measurement results to analyze.}

\item{.penalty}{A numerical penalty or \code{barrett} for penalty computation from \href{https://doi.org/10.1145/3133876}{Barrett et al.}.}
}
\value{
Vector of segment starting points.
}
\description{
Treats samples as segments separated by change points.
Returns indices of the first sample of each segment.
}
\details{
Relies on the PELT algorithm implementation in the \code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}} function.
}
