% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_default_labels}
\alias{plot_default_labels}
\title{Prepare default plot labels.}
\usage{
plot_default_labels(data)
}
\arguments{
\item{data}{Data to generate labels for.}
}
\value{
Tibble with default plot labels.
}
\description{
Prepare default plot labels.
}
