% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_website_violins}
\alias{plot_website_violins}
\title{Plot violin image used on benchmark website.}
\usage{
plot_website_violins(input = ".", warmup = 5 * 60, labels = "labels.csv")
}
\arguments{
\item{input}{Data path or data tibble to summarize.}

\item{warmup}{Duration to discard as warm up time.}

\item{labels}{Labels file or labels tibble.}
}
\description{
Plot violin image used on benchmark website.
}
\examples{
data <- load_path_json (rren_example ())
# Edit labels manually to provide JDK version.
labels <- plot_default_labels (data)
plot_website_violins (data, labels = labels)

}
