% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{preserve_last_n}
\alias{preserve_last_n}
\title{Preserve last samples in each run by count.}
\usage{
preserve_last_n(.data, .n, .max_share = 0.5)
}
\arguments{
\item{.data}{measurement results to filter}

\item{.n}{how many samples taken from end to preserve}

\item{.max_share}{maximum share of samples to preserve}
}
\value{
filtered measurement results
}
\description{
Useful for extracting a set of representative samples for each run,
assuming the last part of each run is the most representative.
Can specify a maximum share of samples that the function
preserves. Runs with fewer samples are removed entirely.
}
