% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{preserve_last_sec}
\alias{preserve_last_sec}
\title{Preserve last samples in each run by time.}
\usage{
preserve_last_sec(.input, .sec, .max_share = 0.5, .equal_count_per_run = FALSE)
}
\arguments{
\item{.input}{Measurement results to filter.}

\item{.sec}{How old samples taken from end to preserve.}

\item{.max_share}{Maximum share of samples to preserve.}

\item{.equal_count_per_run}{Make sample count per run equal.}
}
\value{
Filtered measurement results.
}
\description{
Useful for extracting a set of representative samples for each run,
assuming the last part of each run is the most representative.
Can specify a maximum share of samples that the function
preserves. Runs with fewer samples are removed entirely.
}
\details{
When equal sample count per run is requested,
it is chosen to be the median sample count.
}
