% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{remove_outliers_global}
\alias{remove_outliers_global}
\title{Remove outliers.}
\usage{
remove_outliers_global(.data, .column, ...)
}
\arguments{
\item{.data}{data}

\item{.column}{column to identify outliers in}

\item{...}{parameters to \code{\link[=identify_outliers_global]{identify_outliers_global()}}}
}
\value{
tibble with rows filtered
}
\description{
Uses \code{\link[=identify_outliers_global]{identify_outliers_global()}} to identify outliers in given column. Then, removes the outlier rows.
}
