% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{remove_outliers_window}
\alias{remove_outliers_window}
\title{Remove outliers using sliding window.}
\usage{
remove_outliers_window(.input, .column, ...)
}
\arguments{
\item{.input}{Data.}

\item{.column}{Column to identify outliers in.}

\item{...}{Parameters to \code{\link[=identify_vector_outliers_window]{identify_vector_outliers_window()}}.}
}
\value{
Tibble with rows filtered.
}
\description{
Uses \code{\link[=identify_vector_outliers_window]{identify_vector_outliers_window()}} to identify outliers in given column. Then, removes the outlier rows.
}
