% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_summaries}
\alias{report_summaries}
\title{Report result summaries.}
\usage{
report_summaries(input = ".", warmup = 5 * 60)
}
\arguments{
\item{input}{Data path or data tibble to summarize.}

\item{warmup}{Duration to discard as warm up time.}
}
\description{
Generates result summaries for human consumption.
}
