% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_summaries_benchmark_group_helper}
\alias{report_summaries_benchmark_group_helper}
\title{Group helper for report result summaries.}
\usage{
report_summaries_benchmark_group_helper(group, key)
}
\arguments{
\item{group}{Single benchmark group to report on.}

\item{key}{Group key information.}
}
\description{
Group helper for report result summaries.
}
