% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rren-package.R
\docType{package}
\name{rren-package}
\alias{rren}
\alias{rren-package}
\title{rren: Renaissance Benchmark Data Processing Package}
\description{
A utility package for processing the data produced by the Renaissance benchmark suite (https://renaissance.dev).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/renaissance-benchmarks/utilities-r}
  \item Report bugs at \url{https://github.com/renaissance-benchmarks/utilities-r/issues}
}

}
\author{
\strong{Maintainer}: Petr Tuma \email{petr.tuma@d3s.mff.cuni.cz} (\href{https://orcid.org/0000-0002-7035-2322}{ORCID})

}
\keyword{internal}
