% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{rren_artificial}
\alias{rren_artificial}
\title{Return data frame with artificial data.}
\usage{
rren_artificial(data, vm_count = 1L, run_count = 1L, benchmark_count = 1L)
}
\arguments{
\item{data}{Data to use.}

\item{vm_count}{Count of VM combinations to generate.}

\item{run_count}{Count of run combinations to generate.}

\item{benchmark_count}{Count of benchmark combinations to generate.}
}
\value{
data Frame with artificial data.
}
\description{
Return data frame with artificial data.
}
