% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{rren_example}
\alias{rren_example}
\title{Return full path to an example file.}
\usage{
rren_example(file_name = "")
}
\arguments{
\item{file_name}{Example file name.}
}
\value{
Example file full path.
}
\description{
Return full path to an example file.
}
