% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{check_renaissance}
\alias{check_renaissance}
\title{Check whether data resembles typical measurement results.}
\usage{
check_renaissance(
  .input,
  .check_index = TRUE,
  .check_total = TRUE,
  .check_metadata = TRUE
)
}
\arguments{
\item{.input}{Measurement results to test.}

\item{.check_index}{Check presence of index.}

\item{.check_total}{Check presence of total.}

\item{.check_metadata}{Check presence of descriptive metadata columns.}
}
\value{
Error message or TRUE.
}
\description{
Check whether data resembles typical measurement results.
}
