% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{expect_renaissance}
\alias{expect_renaissance}
\title{Check whether data resembles typical measurement results.}
\usage{
expect_renaissance(.input, info = NULL, label = vname(.data), ...)
}
\arguments{
\item{.input}{Measurement results to test.}

\item{info}{Internal.}

\item{label}{Internal.}

\item{...}{Parameters to \code{\link[=check_renaissance]{check_renaissance()}}.}
}
\value{
An expectation.
}
\description{
Check whether data resembles typical measurement results.
}
\seealso{
\code{\link[=check_renaissance]{check_renaissance()}}
}
