% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{list_outliers_global}
\alias{list_outliers_global}
\title{List outliers.}
\usage{
list_outliers_global(.input, .column, ...)
}
\arguments{
\item{.input}{Data.}

\item{.column}{Column to identify outliers in.}

\item{...}{Parameters to \code{\link[=identify_vector_outliers_global]{identify_vector_outliers_global()}}.}
}
\value{
Tibble with outlier list.
}
\description{
Uses \code{\link[=identify_vector_outliers_global]{identify_vector_outliers_global()}} to identify outliers in given column. Then, returns the list of outlier positions and intervals.
}
