% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{remove_outliers_window}
\alias{remove_outliers_window}
\title{Remove outliers using sliding window.}
\usage{
remove_outliers_window(.data, .column, ...)
}
\arguments{
\item{.data}{data}

\item{.column}{column to identify outliers in}

\item{...}{parameters to \code{\link[=identify_outliers_window]{identify_outliers_window()}}}
}
\value{
tibble with rows filtered
}
\description{
Uses \code{\link[=identify_outliers_window]{identify_outliers_window()}} to identify outliers in given column. Then, removes the outlier rows.
}
