% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{rren_artificial}
\alias{rren_artificial}
\title{Return data frame with artificial data.}
\usage{
rren_artificial(data, run_count = 1L, benchmark_count = 1L, vm_count = 1L)
}
\arguments{
\item{data}{Data to use for each run.}

\item{run_count}{Count of run combinations to generate per virtual machine and benchmark.}

\item{benchmark_count}{Count of benchmark combinations to generate per virtual machine.}

\item{vm_count}{Count of virtual machine combinations to generate.}
}
\value{
data Tibble with artificial data.
}
\description{
Data to use for each run is the same. The result tibble
is arranged to first iterate runs, then benchmarks,
then virtual machines. This guarantee is useful
when difference between runs is to be
injected later.
}
